added a little readlines function which returns a list of lines in a file. renamed all obj_double_t to obj_float_t.

+++ src/parser.tab.h
@@ -47,7 +47,7 @@ extern int yydebug;
       know about them.  */
    enum yytokentype {
      INT = 258,
-     DOUBLE = 259,
+     FLOAT = 259,
      STRING = 260,
      ID = 261,
      NEWLINE = 262,
